+++ CHANGELOG.md
@@ -5,10 +5,71 @@ CHANGELOG
 
 First release of the Platinium Symfony Bundle
 
-### 2.0 ###
+### v1.1.0 ###
 
-Updated for use with
-* Php >=7.1
-* Symfony >= 4.0
+Add langNotIn parameter
 
-Delete Controller and view and change bundle to fonctionnal bundle
+### v1.1.1 ###
+
+Notifer correction for langNotIn, readme update
+
+### v1.1.2 ###
+
+service injection correction
+
+### v1.1.3 ###
+
+change notifier service to add some push options
+
+### v1.1.4 ###
+
+update readme
+Correction newsstand false/true -> 0/1
+
+### v1.1.5 ###
+
+Correction platinium header name
+improve exception message, comments, update composer.json description and add ext php,
+remove unused var in Client and improve parseHttpHeaders method
+
+### v1.1.6 ###
+
+Update required fields from platinium response
+
+### v1.2.0 ###
+
+Update project to Symfony 4.4
+And set minimum PHP version to 7.2
+
+### v1.3.0 ###
+
+Update project to Symfony 5.0
+
+### v1.3.1 ### 
+
+Update project to Symfony 5.1
+
+### v1.3.2 ### 
+
+Update project to Symfony 5.2
+
+### v1.3.3 ### 
+
+Update project to Symfony 5.3
+
+### v1.3.4 ### 
+
+Update project to Symfony 5.4
+
+### v1.4.0 ###
+
+Update project to Symfony 6
+And set minimum PHP version to 8.1
+
+### v1.4.1 ### 
+
+Fix types and simplify PlatiniumPushInformation. Also make tolerance nullable
+
+### v1.4.2 ### 
+
+Use Symfony HttpClient instead of CURL